+++ cli/templates/status.go
@@ -1,17 +1,18 @@
+// Package templates contains all text-ui templates for the command-line
+// interface.
 package templates
 
 // PrettyStatus contains the detailed template
 // for displaying the status of multiple analytics
 // accounts.
-var PrettyStatus = `Global status: {{.OverallStatus}}
+var PrettyStatus = `Known spam domains: {{.KnownSpamDomains}}
 
-Account status:
 {{range .Accounts -}}
-{{printf "%s (%s)" .AccountName .AccountID | printf "%30s"}}: {{.Status}}
+{{printf "%s (%s)" .AccountName .AccountID | printf "%30s"}}: {{printf "%5s" .Status}} {{printf "%d/%d" .Status.UpToDateFilters .Status.TotalFilters | printf "%8s"}}
 {{end}}`
 
 // QuietStatus contains a minimal template for
 // displaying the status of multiple analytics accounts.
 var QuietStatus = `{{range .Accounts -}}
-{{printf "%-10s" .AccountID}} {{printf "%s" .Status}}
+{{printf "%-10s" .AccountID}} {{.Status}}
 {{end}}`